format code, and update capacity in double instead of int

+++ src/main/java/fr/m2/miage/flights/Main.java
@@ -18,7 +18,7 @@ import static fr.m2.miage.flights.services.HibernateSessionProvider.getSessionFa
 public class Main {
     private static final Gson gson = new GsonBuilder().create();
 
-    public static ArrayList<VolAssociation> getVols(TypeVol typeVol, String date, String pays, int capaciteLibre) {
+    public static ArrayList<VolAssociation> getVols(TypeVol typeVol, String date, String pays, double capaciteLibre) {
 
         String query = "Vol.getVolsCorrespondantsALaDemande";
         ArrayList<VolAssociation> volsPourLesAssociation;
@@ -39,7 +39,7 @@ public class Main {
         return volsPourLesAssociation;
     }
 
-    public static List<Object[]> daoGetvols(TypeVol typeVol, String query, String date, String pays, int capaciteLibre) {
+    public static List<Object[]> daoGetvols(TypeVol typeVol, String query, String date, String pays, double capaciteLibre) {
         Session session = getSessionFactory().openSession();
 
         Query queryVolsReguliersCorrespondantsALaDemande = session.createNamedQuery(query, Object[].class);
